Make gate G1 pass, against a real game
35 frames from a live x64 session, replayed through the Rust reader and compared
to what the Electron reader made of the same bytes. No differences.

It took eleven fixes to get there, and every one of them was a bug the gate
existed to find. The gate could not find any of them until now, because the
harness could not replay a frame at all.

Three of those were the harness. The recording carried no offsets bundle, so the
replay tried to rescan -- but the pattern scan runs inside memoryjs and never
passes through a recorded read, so the module's bytes are not in the file and
never will be. It now carries the resolved bundle, once per file. The recorder
wrote only leaf reads and not the pointer derefs along each chain, so a replay
could not take the first step; `offsetAddress` and the .NET string length now
record too. And `SparseProcess` picked the region with the nearest start rather
than one that contains the read, which matters because those new regions overlap
the player record they sit inside: an eight-byte pointer read at record+0x70
hid the byte at record+0x78, and `isDead` came back false for a player the game
had killed.

A fourth was mine twice over: one unparseable address discarded the whole frame,
and a NaN address got into the file in the first place, because a chain step
whose pointer came back undefined makes the next address NaN and `NaN.toString(16)`
is the string "NaN". The recorder refuses those now, and the harness drops a bad
region instead of the frame around it.

The rest were the reader:

- `read_player` looked for `objectPtr`, `taskPtr`, `outfitsPtr`, `rolePtr`,
  `disconnected`, `dead` and `id` in `offsets.player`, where they have never
  been. They are entries of `offsets.player.struct`. Every record failed at the
  first line, so the player list was always empty. `struct_member` lays the
  struct out the way structron does, and the arithmetic checks out: the last
  member ends at exactly the declared `bufferLength`, on both the x86 and x64
  bundles.
- Six top-level fields and all four position fields are chains, read as single
  offsets -- which lands on a pointer instead of the value. Positions are what
  proximity chat is.
- `GameState::from_game` mapped the game's raw state number straight onto this
  enum. They overlap for 0 to 3 while meaning different things, so every lobby
  frame was reported as `Tasks`. The real derivation is a switch, and whether an
  in-game frame is a discussion is a second reading, of the meeting hud. The enum
  decoder that shared the name is now `from_repr`.
- Positions are rounded to four decimal places by the Electron reader, and a
  `f32` cannot hold the result: 36.676 came back as 36.67599868774414. They are
  doubles now, as the collider's `Vector2` already was.
- `mod` is the field's name on the wire, not `modId`.
- `petId` was `data.pet ?? ''` in a field declared `number`. Fixed on the
  TypeScript side, because the type was what was wrong.

What it covers is narrow and worth saying plainly: one map, one game state,
five players, 35 frames. Lobby on The Skeld, 64-bit. Tasks, meetings, vents,
cameras, sabotage and deaths are all still unmeasured, and issue #10 stays open
until they are.

+++ crates/acl-game/src/systems.rs
@@ -40,10 +40,10 @@ const MAX_DOORS: i32 = 16;
 const MIRA_CONSOLES_FOR_CLEAR: u32 = 2;
 
 /// The Skeld's camera console is one place, so proximity to it is the test.
-const SKELD_CAMERA: (f32, f32) = (-12.9364, -2.7928);
+const SKELD_CAMERA: (f64, f64) = (-12.9364, -2.7928);
 
 /// How close the local player must be to count as at The Skeld's camera console.
-const SKELD_CAMERA_RANGE: f32 = 0.6;
+const SKELD_CAMERA_RANGE: f64 = 0.6;
 
 /// The Polus and Airship camera minigames have six cameras; anything else is a bad read.
 const SURVEILLANCE_CAMERAS: u32 = 6;
@@ -123,7 +123,7 @@ pub fn read_systems(
     ship: u64,
     minigame: u64,
     map: MapType,
-    local_position: Option<(f32, f32)>,
+    local_position: Option<(f64, f64)>,
     offsets: &SystemOffsets,
 ) -> Systems {
     let mut systems = Systems::default();
@@ -197,7 +197,7 @@ fn read_camera(
     memory: &dyn ProcessMemory,
     minigame: u64,
     map: MapType,
-    local_position: Option<(f32, f32)>,
+    local_position: Option<(f64, f64)>,
     offsets: &SystemOffsets,
     into: &mut Systems,
 ) {
